requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
email-validator==2.1.1
